small attempt to help scale jupyter for mass consumption
I've moved the majority of the code out of the notebook so that voila
doesn't need to communicate across the websocket connection it
establishes more than once. for normal jupyter notebook usage the
interaction isn't costly but when using it as an interactive dashboard
we want to minimize the number of round trips.

Images are now expected to be in a cdn and some more data processing is
moved to the container build.